# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vibe Sport API",
        "description": "Sports Facility Booking API - MVP",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "VenueListItemDto": {
                "type": "object",
                "properties": {
                    "venueId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sân ABC"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Street, District"
                    },
                    "lat": {
                        "type": "number",
                        "example": 10.762622
                    },
                    "lng": {
                        "type": "number",
                        "example": 106.660172
                    },
                    "distanceKm": {
                        "type": "number",
                        "example": 2.4
                    },
                    "sportTypes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "FOOTBALL"
                        ]
                    }
                },
                "required": [
                    "venueId",
                    "name",
                    "address",
                    "lat",
                    "lng",
                    "sportTypes"
                ]
            },
            "VenueListResponseDto": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VenueListItemDto"
                        }
                    },
                    "page": {
                        "type": "number",
                        "example": 0
                    },
                    "size": {
                        "type": "number",
                        "example": 10
                    },
                    "total": {
                        "type": "number",
                        "example": 1
                    }
                },
                "required": [
                    "items",
                    "page",
                    "size",
                    "total"
                ]
            },
            "CourtItemDto": {
                "type": "object",
                "properties": {
                    "courtId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sân 1"
                    },
                    "sportType": {
                        "type": "string",
                        "example": "FOOTBALL"
                    }
                },
                "required": [
                    "courtId",
                    "name",
                    "sportType"
                ]
            },
            "VenueDetailResponseDto": {
                "type": "object",
                "properties": {
                    "venueId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sân ABC"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Street, District"
                    },
                    "lat": {
                        "type": "number",
                        "example": 10.762622
                    },
                    "lng": {
                        "type": "number",
                        "example": 106.660172
                    },
                    "courts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CourtItemDto"
                        }
                    }
                },
                "required": [
                    "venueId",
                    "name",
                    "address",
                    "lat",
                    "lng",
                    "courts"
                ]
            },
...
            "SlotDto": {
                "type": "object",
                "properties": {
                    "startTime": {
                        "type": "string",
                        "example": "2026-01-10T18:00:00Z"
                    },
                    "endTime": {
                        "type": "string",
                        "example": "2026-01-10T19:00:00Z"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "AVAILABLE",
                            "UNAVAILABLE"
                        ],
                        "example": "AVAILABLE"
                    },
                    "price": {
                        "type": "number",
                        "example": 200000
                    }
                },
                "required": [
                    "startTime",
                    "endTime",
                    "status"
                ]
            },
            "AvailabilityResponseDto": {
                "type": "object",
                "properties": {
                    "courtId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "date": {
                        "type": "string",
                        "example": "2026-01-10"
                    },
                    "slots": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SlotDto"
                        }
                    }
                },
                "required": [
                    "courtId",
                    "date",
                    "slots"
                ]
            },
...
            "CancelBookingDto": {
                "type": "object",
                "properties": {
                    "reason": {
                        "type": "string",
                        "example": "Cannot attend",
                        "description": "Cancellation reason"
                    }
                },
                "required": [
                    "reason"
                ]
            },
            "CancelResponseDto": {
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "CANCELLED_BY_USER",
                            "CANCELLED_BY_OWNER"
                        ],
                        "example": "CANCELLED_BY_USER"
                    }
                },
                "required": [
                    "bookingId",
                    "status"
                ]
            },
...
            "VenueAnalyticsResponseDto": {
                "type": "object",
                "properties": {
                    "totalRevenue": {
                        "type": "number",
                        "example": 1500000
                    },
                    "totalBookings": {
                        "type": "number",
                        "example": 25
                    },
                    "completedBookings": {
                        "type": "number",
                        "example": 20
                    },
                    "cancelledBookings": {
                        "type": "number",
                        "example": 5
                    }
                },
                "required": [
                    "totalRevenue",
                    "totalBookings",
                    "completedBookings",
                    "cancelledBookings"
                ]
            },
...
            "OwnerBookingItemDto": {
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "status": {
                        "type": "string"
                    },
                    "courtId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "startTime": {
                        "type": "string",
                        "example": "2026-01-10T18:00:00Z"
                    },
                    "endTime": {
                        "type": "string",
                        "example": "2026-01-10T19:00:00Z"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerInfoDto"
                    }
                },
                "required": [
                    "bookingId",
                    "status",
                    "courtId",
                    "startTime",
                    "endTime",
                    "customer"
                ]
            },
            "CustomerInfoDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "fullName": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0912345678"
                    }
                },
                "required": [
                    "userId",
                    "fullName"
                ]
            },
            "OwnerBookingsListResponseDto": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OwnerBookingItemDto"
                        }
                    },
                    "page": {
                        "type": "number",
                        "example": 0
                    },
                    "size": {
                        "type": "number",
                        "example": 10
                    },
                    "total": {
                        "type": "number",
                        "example": 1
                    }
                },
                "required": [
                    "items",
                    "page",
                    "size",
                    "total"
                ]
            },
            "OwnerCancelBookingDto": {
                "type": "object",
                "properties": {
                    "reason": {
                        "type": "string",
                        "example": "Court maintenance"
                    }
                },
                "required": [
                    "reason"
                ]
            }
        }
    },
    "paths": {
...
        "/venues": {
            "get": {
                "operationId": "VenuesController_searchVenues",
                "summary": "Search venues",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 10.762622
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 106.660172
                        }
                    },
                    {
                        "name": "radiusKm",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 10,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "sportType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "FOOTBALL"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "sân"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 0,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 10,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VenueListResponseDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Venues"
                ]
            }
        },
        "/venues/{venueId}": {
            "get": {
                "operationId": "VenuesController_getVenueDetail",
                "summary": "Get venue detail with courts",
                "parameters": [
                    {
                        "name": "venueId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VenueDetailResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Venue not found"
                    }
                },
                "tags": [
                    "Venues"
                ]
            }
        },
        "/courts/{courtId}/availability": {
            "get": {
                "operationId": "AvailabilityController_getAvailability",
                "summary": "Get court availability for a specific date",
                "parameters": [
                    {
                        "name": "courtId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "2026-01-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailabilityResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Court not found"
                    }
                },
                "tags": [
                    "Availability"
                ]
            }
        },
...
        "/bookings/{bookingId}/cancel": {
            "post": {
                "operationId": "BookingsController_cancelBooking",
                "summary": "Cancel a booking (customer)",
                "parameters": [
                    {
                        "name": "bookingId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CancelBookingDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CancelResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid transition"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "tags": [
                    "Bookings"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/owner/analytics/{venueId}": {
            "get": {
                "operationId": "OwnerBookingsController_getVenueAnalytics",
                "summary": "Get venue analytics",
                "parameters": [
                    {
                        "name": "venueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VenueAnalyticsResponseDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Owner"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/owner/bookings": {
            "get": {
                "operationId": "OwnerBookingsController_getOwnerBookings",
                "summary": "Get owner bookings",
                "parameters": [
                    {
                        "name": "courtId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnerBookingsListResponseDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Owner"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/owner/bookings/{bookingId}/cancel": {
            "post": {
                "operationId": "OwnerBookingsController_cancelBooking",
                "summary": "Cancel a booking (owner)",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OwnerCancelBookingDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CancelResponseDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Owner"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}